Got markdown parsing working.
Right now it's generating on server startup, but it
seems to me that it might be better to generate at build
time just before waring up the file so that the classpath
won't be an issue.  Plus it's kind of weird to be writing
into the resources directory at runtime.  Either way,
it kinda-sorta works now.